+++ changelog.md
@@ -4,6 +4,8 @@ Changelog
 1.11
 ----
 * Admins can now see all courses from the homepage
+* Admins can now see enrolment information for a student
+* Admins can manually mark a course as complete
 * Fixed display issue with no courses in admin dashboard
 * Fixed bug with most recent lesson not included in progress calculation